+++ app/_mocks/session.ts
@@ -1,5 +1,5 @@
-import { getSessionUserId } from "@/_shared/utils/session";
+import { getSessionUserId } from "@/_shared/utils/getSessionUser";
 
-jest.mock("@/_shared/utils/session");
+jest.mock("@/_shared/utils/getSessionUser");
 export const mockedGetSessionUserId = jest.mocked(getSessionUserId);
 // export const mockedGetSessionUser = jest.mocked(getSessionUser);
