+++ app/page.tsx
@@ -1,6 +1,6 @@
 import { NoteForm } from "./_shared/components/note/NoteForm";
 import { Timeline } from "./_shared/components/note/Timeline";
-import { getSessionUserId } from "./_shared/utils/session";
+import { getSessionUserId } from "./_shared/utils/getSessionUser";
 
 export default async function Page() {
   const userId = await getSessionUserId();
