+++ app/settings/_components/IconSettingsForm/index.tsx
@@ -1,4 +1,4 @@
-import { getSessionUser } from "@/_shared/utils/session";
+import { getSessionUser } from "@/_shared/utils/getSessionUser";
 
 import { IconFileInput } from "./IconFileInput";
