+++ app/users/[userKey]/_components/UserCard/index.tsx
@@ -5,7 +5,7 @@ import { UserIcon } from "@/_shared/components/user/UserIcon";
 import { userService } from "@/_shared/service/user";
 import { env } from "@/_shared/utils/env";
 import { fullUsername } from "@/_shared/utils/fullUsername";
-import { getSessionUserId } from "@/_shared/utils/session";
+import { getSessionUserId } from "@/_shared/utils/getSessionUser";
 
 import { FollowButton } from "./FollowButton";
 import { followCountService } from "./followCountService";
